delete feature attempt 2.0

+++ restful-web-services/src/main/java/com/sept/rest/webservices/restfulwebservices/repositories/PostsRepository.java
@@ -13,7 +13,7 @@ public interface PostsRepository extends JpaRepository<Posts, Long> {
 
 	List<Posts> findByOwnerIdAndDeletedFalse(int studentID);
 
-	Posts findBypostIdAndDeletedFalse(int id);
+	Posts findBypostIdAndDeletedFalse(long id);
 	
 	
 }
